moved the benchmakr page to landing page
